fix: handle observable field correctly
